Refactor test cases to use constants for magic numbers
- Introduced constants for various test parameters in `test_rx_encoder.c`, `test_rx_frame.c`, `test_rx_nanopb.c`, and `test_uart_hal.c` to improve readability and maintainability.
- Replaced hardcoded values with defined constants for counter values, buffer sizes, payload lengths, and UART configurations.
- Ensured consistency across test cases by using these constants, making it easier to update values in the future.

+++ star-rx72n-firmware/lib/rx_core/inc/rx_pin_interface.h
@@ -149,11 +149,11 @@ typedef bool (*rx_pin_is_reserved_fn)(void* ctx, uint8_t port, uint8_t pin);
  *         k_rx_err_gpio_invalid_port if port invalid,
  *         k_rx_err_gpio_invalid_pin if pin invalid
  */
-typedef rx_err_t (*rx_pin_get_function_fn)(void*   ctx,
-                                           uint8_t port,
-                                           uint8_t pin,
-                                           char*   function_out,
-                                           size_t  function_len);
+typedef rx_err_t (*rx_pin_get_function_fn)(void*    ctx,
+                                           uint8_t  port,
+                                           uint8_t  pin,
+                                           char*    function_out,
+                                           uint32_t function_len);
 
 /**
  * @brief Clear all pin reservations
